+++ elasticpress-on-opensearch.php
@@ -14,3 +14,13 @@ add_filter(
 		return '7.10';
 	}
 );
+
+add_filter(
+	'ep_elasticsearch_plugins',
+	function($info) {
+		if (!array_key_exists('ingest-attachment', $info)) {
+			$info['ingest-attachment'] = '2.13.0';
+		}
+		return $info;
+	}
+);